fix(wordcloud): rename wordcloud2.js to .js
